requirements.txt
django
mysqlclient
mongoengine
stomp.py
